#50 Configurar metadatos genericos como plantilla
