+++ src/app/[lang]/call-for-proposals/page.js
@@ -4,6 +4,10 @@ import CallSpeakers from '@/app/[lang]/components/commons/CallSpeakers';
 import Title from './components/Title';
 import Content from './components/Content';
 
+export const metadata = {
+  title: 'Call for proposals'
+};
+
 const callForProposals = () => {
   return (
     <>
