+++ src/app/[lang]/components/FooterLayout.js
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { useState } from 'react';
 import Link from 'next/link';
 import Container from 'react-bootstrap/Container';
