+++ src/app/[lang]/components/alert/News.js
@@ -1,3 +1,5 @@
+'use client';
+
 import React from 'react';
 import Alert from 'react-bootstrap/Alert';
