+++ src/styles/partials/_call-for-proposals.sass
@@ -12,7 +12,7 @@
         .text
             font-size: 23px
             line-height: 35px
-            font-weight: 300
+            font-weight: 400
 
         .closes-date
             border: 1px solid variables.$pink
