+++ src/styles/partials/_keynotes.sass
@@ -7,7 +7,7 @@
   .keynotes-bg
     background-image: url("/images/keynote/background.svg")
     background-size: inherit
-    background-repeat: no-repeat
+    background-repeat: repeat-y
     font-size: 20px
 
     .welcome
@@ -17,7 +17,7 @@
       .text
         font-size: 23px
         line-height: 35px
-        font-weight: 300
+        font-weight: 400
 
     .keynote-card
       align-items: center
